src/bot/requirements.txt
# Базовый состав
aiogram==3.22.0
python-dotenv==1.1.1

# Управление настройками проекта
pydantic-settings==2.10.1

# Логирование
loguru==0.7.3